+++ src/Resources/contao/languages/de/tl_module.php
@@ -1,28 +1,34 @@
 <?php
 
-$arrLang = &$GLOBALS['TL_LANG']['tl_module'];
+$lang = &$GLOBALS['TL_LANG']['tl_module'];
 
 /**
  * Fields
  */
-$arrLang['addManualPagination'][0]         = 'Manuelle Paginierung hinzufügen';
-$arrLang['addManualPagination'][1]         = 'Wählen Sie diese Option, um dem Modul eine manuelle Paginierung über Inhaltselemente hinzuzufügen.';
-$arrLang['addPagination'][0]               = 'Paginierung hinzufügen';
-$arrLang['addPagination'][1]               = 'Wählen Sie diese Option, um dem Modul eine Paginierung hinzuzufügen.';
-$arrLang['paginationMaxCharCount'][0]      = 'Maximale Zeichenanzahl';
-$arrLang['paginationMaxCharCount'][1]      = 'Geben Sie hier an, wie viele Zeichen ein Abschnitt der Paginierung enthalten soll.';
-$arrLang['paginationCeTextCssSelector'][0] = 'CSS-Selektor in ce_text';
-$arrLang['paginationCeTextCssSelector'][1] = 'Wenn Sie das Template "ce_text" überschrieben haben, können Sie hier den CSS-Selektor anpassen, der zu den Elementen führt, die den Text enthalten, also bspw. den p-Elementen (Standard: ".ce_text").';
-$arrLang['avoidTrailingHeadlines'][0]      = 'Überschriften am Ende einer Seite vermeiden';
-$arrLang['avoidTrailingHeadlines'][1]      = 'Wählen Sie diese Option, um Überschriften und Absätze zusammenzuhalten.';
-$arrLang['addFullVersionCanonicalLink'][0] = 'Canonical-Link zum Head-Element hinzufügen';
-$arrLang['addFullVersionCanonicalLink'][1] = 'Wählen Sie diese Option, um dem &lt;head&gt;-Element Informationen zur Paginierung hinzuzufügen.';
-$arrLang['fullVersionGetParameter'][0]     = 'GET-Parameter für nichtpaginierte Version der Nachricht';
-$arrLang['fullVersionGetParameter'][1]     = 'Geben Sie hier den Namen des Parameters ein, der vorhanden sein muss, damit die nichtpaginierte Version der Nachricht angezeigt wird. Sinnvoll aus SEO-Gründen und für eine Druckenfunktion.';
-$arrLang['setPrevNextLinks'][0]            = 'Prev- und Next-Links zum Head-Element hinzufügen';
-$arrLang['setPrevNextLinks'][1]            = 'Wählen Sie diese Option, um dem &lt;head&gt;-Element die entsprechenden Relationslinks hinzuzufügen.';
+$lang['paginationMode'][0]              = 'Paginierungsmodus';
+$lang['paginationMode'][1]              = 'Wählen Sie hier einen Paginierungsmodus aus.';
+$lang['paginationMaxCharCount'][0]      = 'Maximale Zeichenanzahl';
+$lang['paginationMaxCharCount'][1]      = 'Geben Sie hier an, wie viele Zeichen ein Abschnitt der Paginierung enthalten soll.';
+$lang['paginationCeTextCssSelector'][0] = 'CSS-Selektor in ce_text';
+$lang['paginationCeTextCssSelector'][1] = 'Wenn Sie das Template "ce_text" überschrieben haben, können Sie hier den CSS-Selektor anpassen, der zu den Elementen führt, die den Text enthalten, also bspw. den p-Elementen (Standard: ".ce_text").';
+$lang['avoidTrailingHeadlines'][0]      = 'Überschriften am Ende einer Seite vermeiden';
+$lang['avoidTrailingHeadlines'][1]      = 'Wählen Sie diese Option, um Überschriften und Absätze zusammenzuhalten.';
+$lang['addFullVersionCanonicalLink'][0] = 'Canonical-Link zum Head-Element hinzufügen';
+$lang['addFullVersionCanonicalLink'][1] = 'Wählen Sie diese Option, um dem &lt;head&gt;-Element Informationen zur Paginierung hinzuzufügen.';
+$lang['fullVersionGetParameter'][0]     = 'GET-Parameter für nichtpaginierte Version der Nachricht';
+$lang['fullVersionGetParameter'][1]     = 'Geben Sie hier den Namen des Parameters ein, der vorhanden sein muss, damit die nichtpaginierte Version der Nachricht angezeigt wird. Sinnvoll aus SEO-Gründen.';
+$lang['acceptPrintGetParameter'][0]     = 'GET-Parameter "print" für nichtpaginierte Version der Nachricht akzeptieren';
+$lang['acceptPrintGetParameter'][1]     = 'Wählen Sie diese Option, damit die nichtpaginierte Version der Nachricht angezeigt wird, wenn der Parameter "print" vorhanden ist. Sinnvoll für eine Druckenfunktion.';
+$lang['setPrevNextLinks'][0]            = 'Prev- und Next-Links zum Head-Element hinzufügen';
+$lang['setPrevNextLinks'][1]            = 'Wählen Sie diese Option, um dem &lt;head&gt;-Element die entsprechenden Relationslinks hinzuzufügen.';
 
 /**
  * Legends
  */
-$arrLang['pagination_legend'] = 'Paginierung';
+$lang['pagination_legend'] = 'Paginierung';
+
+$lang['reference']['newsPaginationBundle'] = [
+    \HeimrichHannot\NewsPaginationBundle\NewsPaginationBundle::MODE_AUTO                      => 'Automatisch',
+    \HeimrichHannot\NewsPaginationBundle\NewsPaginationBundle::MODE_MANUAL                    => 'Manuell',
+    \HeimrichHannot\NewsPaginationBundle\NewsPaginationBundle::MODE_MANUAL_WITH_AUTO_FALLBACK => 'Manuell mit Fallback auf Automatisch',
+];